# api-gateway/OpenApi/openapi.yaml
openapi: 3.0.0
info:
  title: Event Planner API
  version: 1.0.0
paths:
  /usuarios/register: 
    post:
      summary: Registrar usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
      responses:
        '200':
          description: Usuario registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensaje'
        '400':
          description: Error en la solicitud
  /usuarios/login: 
    post:
      summary: Iniciar sesión (Necesario para obtener el token, solo con usuarios registrados)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        '200':
          description: Usuario autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Error en la solicitud
  /usuarios/usuario: 
    get:
      summary: Obtener usuario
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuario obtenido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Error en la solicitud
  /usuarios/actualizar: 
    put:
      summary: Actualizar usuario
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Actualizar'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensaje'
        '400':
          description: Error en la solicitud
...
  /eventos/listaNotificacion: 
    get:
      summary: Obtener lista de notificaciones
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de notificaciones obtenida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notificacion'
        '400':
          description: Error en la solicitud
  /eventos/crear: 
    post:
      summary: Crear evento (Es necesario crear un evento para probar las siguientes funcionalidades)
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mensaje'
      responses:
        '200':
          description: Evento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evento'
        '400':
          description: Error en la solicitud
...
components:
  schemas:
...
    Notificacion:
      type: array
      items:
        type: string
    Actualizar:
      type: object
      properties:
        contrasenaVieja:
          type: string
        contrasenaNueva:
          type: string
        contrasenaRepetida:
          type: string
        correo:
          type: string
      required:
        - contrasenaVieja
        - contrasenaNueva
        - contrasenaRepetida
        - correo
    Register:
      type: object
      properties:
        nombreUsuario:
          type: string
        contrasena:
          type: string
        correo:
          type: string
      required:
        - nombreUsuario
        - contrasena
        - correo
...
    Login:
      type: object
      properties:
        nombreUsuario:
          type: string
        contrasena:
          type: string
      required:
        - nombreUsuario
        - contrasena
    Token:
      type: object
      properties:
        token:
          type: string
      required:
        - token
...
    Usuario:
      type: object
      properties:
        idUsuario:
          type: string
        nombreUsuario:
          type: string
        correo:
          type: string
      required:
        - idUsuario
        - nombreUsuario
        - correo
    Evento:
      type: object
      properties:
        idEvento:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        lugar:
          type: string
        fechaini:
          type: string
        fechafin:
          type: string
        idUsuario:
          type: string
        correo:
          type: string
        invitados:
          type: array
          items:
            type: object
            properties:
              idUsuario:
                type: string
              nombre:
                type: string
              correo:
                type: string
        required:
        - idEvento
        - nombre
        - descripcion
        - lugar
        - fechaini
        - fechafin
        - idUsuario
        - correo
        - invitados